Update parseForm validation

+++ app/_actions/memberLevel.tsx
@@ -5,6 +5,7 @@ import {z} from 'zod';
 import { revalidatePath } from 'next/cache';
 import prisma from '@/prisma/prisma';
 import { parsedEnv } from '../_libs/zod-env';
+import { SelectMemberLevelSchema, InsertMemberLevelSchema, UpdateMemberLevelSchema, DeleteMemberLevelSchema } from '@/app/_libs/zod-form';
 
 const UUID5_NAMESPACE = parsedEnv.UUID5_NAMESPACE;
 const UUID5_SECRET = uuidv5(UUID5_NAMESPACE, uuidv5.DNS)
@@ -19,7 +20,7 @@ export async function getMemberLevel() {
 
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/memberLevel');
 
         return result
@@ -35,16 +36,7 @@ export async function insertMemberLevel(prevState: any, formData: FormData) {
     // Set current datetime
     const now = new Date();
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        member_level_id: z.string().uuid(),
-        member_level_name: z.string(),
-        bonus_points_min: z.coerce.number().nonnegative(),
-        bonus_points_max: z.coerce.number().nonnegative(),
-        member_level_created_date: z.date(),
-        member_level_updated_date: z.date()
-    });
-    const data = schema.parse({
+    const parsedForm = InsertMemberLevelSchema.safeParse({
         member_level_id: uuidv5(formData.get('member_level_name'), UUID5_SECRET),
         member_level_name: formData.get('member_level_name'),
         bonus_points_min: formData.get('bonus_points_min'),
@@ -53,18 +45,22 @@ export async function insertMemberLevel(prevState: any, formData: FormData) {
         member_level_updated_date: now
     });
 
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
+
     try {
 
         // get result from prisma
         const result = await prisma.memberLevel.create({
-            data: data,
+            data: parsedForm.data,
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/memberLevel');
 
 
-        return { message: `Successfully inserted ${data['member_level_id']}` }
+        return { message: `Successfully inserted ${parsedForm.data['member_level_id']}` }
 
     } catch(e) {
         return { message: 'Failed to insert the item' }
@@ -77,16 +73,7 @@ export async function updateMemberLevel(prevState: any, formData: FormData) {
     // Set current datetime
     const now = new Date();
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        member_level_id: z.string().uuid(),
-        member_level_name: z.string(),
-        bonus_points_min: z.coerce.number().nonnegative(),
-        bonus_points_max: z.coerce.number().nonnegative(),
-        member_level_created_date: z.date(),
-        member_level_updated_date: z.date()
-    });
-    const data = schema.parse({
+    const parsedForm = UpdateMemberLevelSchema.safeParse({
         member_level_id: formData.get('member_level_id'),
         member_level_name: formData.get('member_level_name'),
         bonus_points_min: formData.get('bonus_points_min'),
@@ -95,22 +82,26 @@ export async function updateMemberLevel(prevState: any, formData: FormData) {
         member_level_updated_date: now
     });
 
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
+
     try {
 
         // get result from prisma
         const result = await prisma.memberLevel.update({
             where: {
-                member_level_id: data['member_level_id']
+                member_level_id: parsedForm.data['member_level_id']
             },
 
-            data: data,
+            data: parsedForm.data,
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/memberLevel');
 
 
-        return { message: `Successfully updated ${data['member_level_id']}` }
+        return { message: `Successfully updated ${parsedForm.data['member_level_id']}` }
 
     } catch(e) {
         return { message: 'Failed to update the item' }
@@ -120,27 +111,27 @@ export async function updateMemberLevel(prevState: any, formData: FormData) {
 
 export async function deleteMemberLevel(prevState: any, formData: FormData) {
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        member_level_id: z.string().uuid()
-    });
-    const data = schema.parse({
+    const parsedForm = DeleteMemberLevelSchema.safeParse({
         member_level_id: formData.get('member_level_id')
     });
+
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
 
     try {
 
         // get result from prisma
         const result = await prisma.memberLevel.delete({
             where: {
-                member_level_id: data['member_level_id']
+                member_level_id: parsedForm.data['member_level_id']
             },
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/memberLevel');
 
-        return { message: `Successfully deleted ${data['member_level_id']}` }
+        return { message: `Successfully deleted ${parsedForm.data['member_level_id']}` }
 
     } catch(e) {
         return { message: 'Failed to delete the item' }
