+++ app/_actions/order.tsx
@@ -5,6 +5,7 @@ import {z} from 'zod';
 import { revalidatePath } from 'next/cache';
 import prisma from '@/prisma/prisma';
 import { parsedEnv } from '../_libs/zod-env';
+import { SelectOrderSchema, InsertOrderSchema, UpdateOrderSchema, DeleteOrderSchema } from '../_libs/zod-form';
 
 const UUID5_NAMESPACE = parsedEnv.UUID5_NAMESPACE;
 const UUID5_SECRET = uuidv5(UUID5_NAMESPACE, uuidv5.DNS)
@@ -22,7 +23,7 @@ export async function getMember() {
             }
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/');
 
         return result
@@ -47,7 +48,7 @@ export async function getOrder() {
 
         // });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/');
 
         return result
@@ -63,32 +64,29 @@ export async function insertOrder(prevState: any, formData: FormData) {
     // Set current datetime
     const now = new Date();
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        order_id: z.string().uuid(),
-        member_id: z.string().uuid(),
-        order_created_date: z.date(),
-        order_updated_date: z.date()
-    });
-    const data = schema.parse({
+    const parsedForm = InsertOrderSchema.safeParse({
         order_id: uuidv5(formData.get('member_id') + (UUID5_DELIMITER || '|') + now.toString, UUID5_SECRET),
         member_id: formData.get('member_id'),
         order_created_date: now,
         order_updated_date: now
     });
+    
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
 
     try {
 
         // get result from prisma
         const result = await prisma.order.create({
-            data: data,
+            data: parsedForm.data,
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/');
 
 
-        return { message: `Successfully inserted ${data['order_id']}` }
+        return { message: `Successfully inserted ${parsedForm.data['order_id']}` }
 
     } catch(e) {
         return { message: 'Failed to insert the item' }
@@ -101,34 +99,32 @@ export async function updateOrder(prevState: any, formData: FormData) {
     // Set current datetime
     const now = new Date();
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        order_id: z.string().uuid(),
-        member_id: z.string().uuid(),
-        order_updated_date: z.date()
-    });
-    const data = schema.parse({
+    const parsedForm = UpdateOrderSchema.safeParse({
         order_id: formData.get('order_id'),
         member_id: formData.get('member_id'),
         order_updated_date: now
     });
+    
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
 
     try {
 
         // get result from prisma
         const result = await prisma.order.update({
             where: {
-                order_id: data['order_id']
+                order_id: parsedForm.data['order_id']
             },
 
-            data: data,
+            data: parsedForm.data,
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/');
 
 
-        return { message: `Successfully updated ${data['order_id']}` }
+        return { message: `Successfully updated ${parsedForm.data['order_id']}` }
 
     } catch(e) {
         return { message: 'Failed to update the item' }
@@ -138,27 +134,27 @@ export async function updateOrder(prevState: any, formData: FormData) {
 
 export async function deleteOrder(prevState: any, formData: FormData) {
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        order_id: z.string().uuid()
-    });
-    const data = schema.parse({
+    const parsedForm = DeleteOrderSchema.safeParse({
         order_id: formData.get('order_id')
     });
+    
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
 
     try {
 
         // get result from prisma
         const result = await prisma.order.delete({
             where: {
-                order_id: data['order_id']
+                order_id: parsedForm.data['order_id']
             },
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/');
 
-        return { message: `Successfully deleted ${data['order_id']}` }
+        return { message: `Successfully deleted ${parsedForm.data['order_id']}` }
 
     } catch(e) {
         return { message: 'Failed to delete the item' }
