+++ app/_actions/product.tsx
@@ -5,6 +5,7 @@ import {z} from 'zod';
 import { revalidatePath } from 'next/cache';
 import prisma from '@/prisma/prisma';
 import { parsedEnv } from '../_libs/zod-env';
+import { SelectProductSchema, InsertProductSchema, UpdateProductSchema, DeleteProductSchema } from '../_libs/zod-form';
 
 const UUID5_NAMESPACE = parsedEnv.UUID5_NAMESPACE;
 const UUID5_SECRET = uuidv5(UUID5_NAMESPACE, uuidv5.DNS)
@@ -22,7 +23,7 @@ export async function getUOM() {
             }
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/product');
 
         return result
@@ -49,7 +50,7 @@ export async function getProduct() {
         //     }
         // });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/product');
 
         return result
@@ -65,17 +66,7 @@ export async function insertProduct(prevState: any, formData: FormData) {
     // Set current datetime
     const now = new Date();
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        product_id: z.string().uuid(),
-        product_name: z.string(),
-        uom_id: z.string().uuid(),
-        product_unit_price: z.coerce.number().nonnegative(),
-        product_bonus_points: z.coerce.number().nonnegative(),
-        product_created_date: z.date(),
-        product_updated_date: z.date()
-    });
-    const data = schema.parse({
+    const parsedForm = InsertProductSchema.safeParse({
         product_id: uuidv5(formData.get('product_name') + (UUID5_DELIMITER || '|') + formData.get('uom_name'), UUID5_SECRET),
         product_name: formData.get('product_name'),
         uom_id: uuidv5(formData.get('uom_name'), UUID5_SECRET),
@@ -84,19 +75,23 @@ export async function insertProduct(prevState: any, formData: FormData) {
         product_created_date: now,
         product_updated_date: now
     });
+    
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
 
     try {
 
         // get result from prisma
         const result = await prisma.product.create({
-            data: data,
+            data: parsedForm.data,
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/product');
 
 
-        return { message: `Successfully inserted ${data['product_id']}` }
+        return { message: `Successfully inserted ${parsedForm.data['product_id']}` }
 
     } catch(e) {
         return { message: 'Failed to insert the item' }
@@ -109,17 +104,7 @@ export async function updateProduct(prevState: any, formData: FormData) {
     // Set current datetime
     const now = new Date();
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        product_id: z.string().uuid(),
-        product_name: z.string(),
-        uom_id: z.string().uuid(),
-        product_unit_price: z.coerce.number().nonnegative(),
-        product_bonus_points: z.coerce.number().nonnegative(),
-        product_created_date: z.date(),
-        product_updated_date: z.date()
-    });
-    const data = schema.parse({
+    const parsedForm = UpdateProductSchema.safeParse({
         product_id: formData.get('product_id'),
         product_name: formData.get('product_name'),
         uom_id: uuidv5(formData.get('uom_name'), UUID5_SECRET),
@@ -127,23 +112,27 @@ export async function updateProduct(prevState: any, formData: FormData) {
         product_bonus_points: formData.get('product_bonus_points'),
         product_updated_date: now
     });
+    
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
 
     try {
 
         // get result from prisma
         const result = await prisma.product.update({
             where: {
-                product_id: data['product_id']
+                product_id: parsedForm.data['product_id']
             },
 
-            data: data,
+            data: parsedForm.data,
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/product');
 
 
-        return { message: `Successfully updated ${data['product_id']}` }
+        return { message: `Successfully updated ${parsedForm.data['product_id']}` }
 
     } catch(e) {
         return { message: 'Failed to update the item' }
@@ -153,27 +142,27 @@ export async function updateProduct(prevState: any, formData: FormData) {
 
 export async function deleteProduct(prevState: any, formData: FormData) {
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        product_id: z.string().uuid()
-    });
-    const data = schema.parse({
+    const parsedForm = DeleteProductSchema.safeParse({
         product_id: formData.get('product_id')
     });
+    
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
 
     try {
 
         // get result from prisma
         const result = await prisma.product.delete({
             where: {
-                product_id: data['product_id']
+                product_id: parsedForm.data['product_id']
             },
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/product');
 
-        return { message: `Successfully deleted ${data['product_id']}` }
+        return { message: `Successfully deleted ${parsedForm.data['product_id']}` }
 
     } catch(e) {
         return { message: 'Failed to delete the item' }
