+++ app/_actions/uom.tsx
@@ -5,6 +5,7 @@ import {z} from 'zod';
 import { revalidatePath } from 'next/cache';
 import prisma from '@/prisma/prisma';
 import { parsedEnv } from '../_libs/zod-env';
+import { SelectUomSchema, InsertUomSchema, UpdateUomSchema, DeleteUomSchema } from '@/app/_libs/zod-form';
 
 const UUID5_NAMESPACE = parsedEnv.UUID5_NAMESPACE;
 const UUID5_SECRET = uuidv5(UUID5_NAMESPACE, uuidv5.DNS)
@@ -19,7 +20,7 @@ export async function getUOM() {
 
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/uom');
 
         return result
@@ -35,32 +36,29 @@ export async function insertUOM(prevState: any, formData: FormData) {
     // Set current datetime
     const now = new Date();
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        uom_id: z.string().uuid(),
-        uom_name: z.string(),
-        uom_created_date: z.date(),
-        uom_updated_date: z.date()
-    });
-    const data = schema.parse({
+    const parsedForm = InsertUomSchema.safeParse({
         uom_id: uuidv5(formData.get('uom_name'), UUID5_SECRET),
         uom_name: formData.get('uom_name'),
         uom_created_date: now,
         uom_updated_date: now
     });
 
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
+
     try {
 
         // get result from prisma
         const result = await prisma.uom.create({
-            data: data,
+            data: parsedForm.data,
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/uom');
 
 
-        return { message: `Successfully inserted ${data['uom_id']}` }
+        return { message: `Successfully inserted ${parsedForm.data['uom_id']}` }
 
     } catch(e) {
         return { message: 'Failed to insert the item' }
@@ -73,35 +71,32 @@ export async function updateUOM(prevState: any, formData: FormData) {
     // Set current datetime
     const now = new Date();
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        uom_id: z.string().uuid(),
-        uom_name: z.string(),
-        uom_created_date: z.date(),
-        uom_updated_date: z.date()
-    });
-    const data = schema.parse({
+    const parsedForm = UpdateUomSchema.safeParse({
         uom_id: formData.get('uom_name'),
         uom_name: formData.get('uom_name'),
         uom_updated_date: now
     });
+
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
 
     try {
 
         // get result from prisma
         const result = await prisma.uom.update({
             where: {
-                uom_id: data['uom_id']
+                uom_id: parsedForm.data['uom_id']
             },
 
-            data: data,
+            data: parsedForm.data,
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/uom');
 
 
-        return { message: `Successfully updated ${data['uom_id']}` }
+        return { message: `Successfully updated ${parsedForm.data['uom_id']}` }
 
     } catch(e) {
         return { message: 'Failed to update the item' }
@@ -111,27 +106,27 @@ export async function updateUOM(prevState: any, formData: FormData) {
 
 export async function deleteUOM(prevState: any, formData: FormData) {
 
-    // set zod schema to validate form data
-    const schema = z.object({
-        uom_id: z.string().uuid()
-    });
-    const data = schema.parse({
+    const parsedForm = DeleteUomSchema.safeParse({
         uom_id: formData.get('uom_id')
     });
+    
+    if (!parsedForm.success) {
+        return { message: parsedForm.error.toString()};
+    };
 
     try {
 
         // get result from prisma
         const result = await prisma.uom.delete({
             where: {
-                uom_id: data['uom_id']
+                uom_id: parsedForm.data['uom_id']
             },
         });
 
-        // Check existing cache, revalidate with the fetched data
+        // Invalidate existing cache, forcing static site re-rendering
         revalidatePath('/uom');
 
-        return { message: `Successfully deleted ${data['uom_id']}` }
+        return { message: `Successfully deleted ${parsedForm.data['uom_id']}` }
 
     } catch(e) {
         return { message: 'Failed to delete the item' }
